Code review with Nemanja

+++ boards/cats_rev1Pro/src/util/actions.c
@@ -23,26 +23,26 @@
 #include "lfs/lfs_custom.h"
 #include "drivers/servo.h"
 
-bool no_action_function(__attribute__((unused)) int32_t bummer);
+bool no_action_function(__attribute__((unused)) int16_t bummer);
 
-bool os_delay(int32_t ticks);
+bool os_delay(int16_t ticks);
 
-bool high_current_channel_one(int32_t state);
-bool high_current_channel_two(int32_t state);
-bool high_current_channel_three(int32_t state);
-bool high_current_channel_four(int32_t state);  // reserved for later use
-bool high_current_channel_five(int32_t state);  // reserved for later use
-bool high_current_channel_six(int32_t state);   // reserved for later use
+bool high_current_channel_one(int16_t state);
+bool high_current_channel_two(int16_t state);
+bool high_current_channel_three(int16_t state);
+bool high_current_channel_four(int16_t state);  // reserved for later use
+bool high_current_channel_five(int16_t state);  // reserved for later use
+bool high_current_channel_six(int16_t state);   // reserved for later use
 
-bool low_level_channel_one(int32_t state);
-bool low_level_channel_two(int32_t state);
-bool low_level_channel_three(int32_t state);  // reserved for later use
-bool low_level_channel_four(int32_t state);   // reserved for later use
+bool low_level_channel_one(int16_t state);
+bool low_level_channel_two(int16_t state);
+bool low_level_channel_three(int16_t state);  // reserved for later use
+bool low_level_channel_four(int16_t state);   // reserved for later use
 
-bool servo_channel_one(int32_t angle);
-bool servo_channel_two(int32_t angle);
-bool servo_channel_three(int32_t angle);  // reserved for later use
-bool servo_channel_four(int32_t angle);   // reserved for later use
+bool servo_channel_one(int16_t angle);
+bool servo_channel_two(int16_t angle);
+bool servo_channel_three(int16_t angle);  // reserved for later use
+bool servo_channel_four(int16_t angle);   // reserved for later use
 
 const peripheral_act_fp action_table[NUM_ACTION_FUNCTIONS] = {no_action_function,         os_delay,
                                                               high_current_channel_one,   high_current_channel_two,
@@ -54,7 +54,7 @@ const peripheral_act_fp action_table[NUM_ACTION_FUNCTIONS] = {no_action_function
                                                               servo_channel_three,        servo_channel_four,
                                                               set_recorder_state};
 
-bool no_action_function(__attribute__((unused)) int32_t bummer) {
+bool no_action_function(__attribute__((unused)) int16_t bummer) {
   // Sucks to be here...
   // it seems like someone didn't configure the actions right
   return false;
@@ -69,7 +69,7 @@ bool no_action_function(__attribute__((unused)) int32_t bummer) {
  * Right now this is considered a feature and not a bug since we
  * assume the users know what they are doing when setting up these
  * delays. */
-bool os_delay(int32_t ticks) {
+bool os_delay(int16_t ticks) {
   if (ticks > 0) {
     osDelay((uint32_t)ticks);
     return true;
@@ -78,8 +78,7 @@ bool os_delay(int32_t ticks) {
 }
 
 // High current outputs for pyros, valves etc.
-
-bool high_current_channel_one(int32_t state) {
+bool high_current_channel_one(int16_t state) {
   if (state == 0 || state == 1) {
     HAL_GPIO_WritePin(PYRO1_GPIO_Port, PYRO1_Pin, (GPIO_PinState)state);
     return true;
@@ -87,7 +86,7 @@ bool high_current_channel_one(int32_t state) {
   return false;
 }
 
-bool high_current_channel_two(int32_t state) {
+bool high_current_channel_two(int16_t state) {
   if (state == 0 || state == 1) {
     HAL_GPIO_WritePin(PYRO2_GPIO_Port, PYRO2_Pin, (GPIO_PinState)state);
     return true;
@@ -95,7 +94,7 @@ bool high_current_channel_two(int32_t state) {
   return false;
 }
 
-bool high_current_channel_three(int32_t state) {
+bool high_current_channel_three(int16_t state) {
   if (state == 0 || state == 1) {
     HAL_GPIO_WritePin(PYRO3_GPIO_Port, PYRO3_Pin, (GPIO_PinState)state);
     return true;
@@ -103,7 +102,7 @@ bool high_current_channel_three(int32_t state) {
   return false;
 }
 
-bool high_current_channel_four(int32_t state) {
+bool high_current_channel_four(int16_t state) {
   if (state == 0 || state == 1) {
     // HAL_GPIO_WritePin(PYRO_3_GPIO_Port, PYRO_3_Pin,  (GPIO_PinState)state);
     return true;
@@ -111,7 +110,7 @@ bool high_current_channel_four(int32_t state) {
   return false;
 }
 
-bool high_current_channel_five(int32_t state) {
+bool high_current_channel_five(int16_t state) {
   if (state == 0 || state == 1) {
     // HAL_GPIO_WritePin(PYRO_3_GPIO_Port, PYRO_3_Pin,  (GPIO_PinState)state);
     return true;
@@ -119,7 +118,7 @@ bool high_current_channel_five(int32_t state) {
   return false;
 }
 
-bool high_current_channel_six(int32_t state) {
+bool high_current_channel_six(int16_t state) {
   if (state == 0 || state == 1) {
     // HAL_GPIO_WritePin(PYRO_3_GPIO_Port, PYRO_3_Pin,  (GPIO_PinState)state);
     return true;
@@ -129,7 +128,7 @@ bool high_current_channel_six(int32_t state) {
 
 // Low level (3.3V) outputs
 
-bool low_level_channel_one(int32_t state) {
+bool low_level_channel_one(int16_t state) {
   if (state == 0 || state == 1) {
     HAL_GPIO_WritePin(IO1_GPIO_Port, IO1_Pin, (GPIO_PinState)state);
     return true;
@@ -137,7 +136,7 @@ bool low_level_channel_one(int32_t state) {
   return false;
 }
 
-bool low_level_channel_two(int32_t state) {
+bool low_level_channel_two(int16_t state) {
   if (state == 0 || state == 1) {
     HAL_GPIO_WritePin(IO2_GPIO_Port, IO2_Pin, (GPIO_PinState)state);
     return true;
@@ -146,7 +145,7 @@ bool low_level_channel_two(int32_t state) {
 }
 
 // Same as servo 1 but digital output
-bool low_level_channel_three(int32_t state) {
+bool low_level_channel_three(int16_t state) {
   if (state == 0 || state == 1) {
     servo_set_onoff(&SERVO1, state);
     return true;
@@ -155,7 +154,7 @@ bool low_level_channel_three(int32_t state) {
 }
 
 // Same as servo 2 but digital output
-bool low_level_channel_four(int32_t state) {
+bool low_level_channel_four(int16_t state) {
   if (state == 0 || state == 1) {
     servo_set_onoff(&SERVO2, (bool)state);
     return true;
@@ -165,7 +164,7 @@ bool low_level_channel_four(int32_t state) {
 
 // Servo Outputs
 
-bool servo_channel_one(int32_t angle) {
+bool servo_channel_one(int16_t angle) {
   if (angle >= 0 && angle <= 180) {
     servo_set_position(&SERVO1, angle);
     return true;
@@ -173,7 +172,7 @@ bool servo_channel_one(int32_t angle) {
   return false;
 }
 
-bool servo_channel_two(int32_t angle) {
+bool servo_channel_two(int16_t angle) {
   if (angle >= 0 && angle <= 180) {
     servo_set_position(&SERVO2, angle);
     return true;
@@ -181,7 +180,7 @@ bool servo_channel_two(int32_t angle) {
   return false;
 }
 
-bool servo_channel_three(int32_t angle) {
+bool servo_channel_three(int16_t angle) {
   if (angle >= 0 && angle <= 180) {
     // servo_set_position(&SERVO1, angle);
     return true;
@@ -189,7 +188,7 @@ bool servo_channel_three(int32_t angle) {
   return false;
 }
 
-bool servo_channel_four(int32_t angle) {
+bool servo_channel_four(int16_t angle) {
   if (angle >= 0 && angle <= 180) {
     // servo_set_position(&SERVO1, angle);
     return true;
@@ -198,7 +197,7 @@ bool servo_channel_four(int32_t angle) {
 }
 
 /* TODO check if mutex should be used here */
-bool set_recorder_state(int32_t state) {
+bool set_recorder_state(int16_t state) {
   volatile recorder_status_e new_rec_state = (recorder_status_e)state;
   if (new_rec_state < REC_OFF || new_rec_state > REC_WRITE_TO_FLASH) {
     return false;
